RUMM-1964 Update `distribution` tool to separate release-related classes
so we can share some of this code with dogfooding tool.
